% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttons.R
\name{DT_button_extend}
\alias{DT_button_extend}
\title{Create a list to configure \href{https://datatables.net/extensions/buttons/}{Datatable Buttons extension} options
\href{https://datatables.net/extensions/buttons/examples/styling/icons.html}{The Icons page} demonstrates the structure of the Buttons extended options}
\usage{
DT_button_extend(extend, text, icon = NULL, class = NULL, ...)
}
\arguments{
\item{extend}{\code{chr} The button type to extend}

\item{text}{\code{chr} Button text}

\item{icon}{\code{chr/shiny.tag} A \code{shiny.tag} generated by \link[shiny]{icon} or a character naming an icon}

\item{class}{\code{chr} Classes to append to the button text or icon}

\item{...}{\code{additional named options}}
}
\value{
\code{list} formatted for consumption by DT
}
\description{
Create a list to configure \href{https://datatables.net/extensions/buttons/}{Datatable Buttons extension} options
\href{https://datatables.net/extensions/buttons/examples/styling/icons.html}{The Icons page} demonstrates the structure of the Buttons extended options
}
\examples{
# An HTML 5 Copy button with the FontAwesome copy icon
datatable_button_extend(
'copyHtml5',
text = "Copy Visible",
icon = "copy"
)
}
