% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{DT_filter_restore}
\alias{DT_filter_restore}
\title{Transforms the DT search_columns input into the \code{searchCols} option list}
\usage{
DT_filter_restore(
  search_columns = session$input[[paste0(outputId, "_search_columns")]],
  outputId,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{search_columns}{\code{chr} The value stored in input$\link{outputId}_search_columns}
}
\value{
\code{list} A list object ready for use in Datatable options argument
}
\description{
\strong{IMPORTANT} Only use on DT where the data will not change!
For an explanation of how this option must be formatted, see this \href{https://github.com/rstudio/DT/issues/77#issuecomment-107257008}{comment}.
}
