% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{DT_formatNum}
\alias{DT_formatNum}
\title{Format numbers in a datatable}
\usage{
DT_formatNum(
  table,
  columns = names(purrr::keep(table$x$data, is.numeric)),
  currency = "",
  interval = 3,
  mark = ",",
  digits = 0,
  dec.mark = getOption("OutDec"),
  before = TRUE,
  zero.print = NULL,
  rows = NULL
)
}
\arguments{
\item{table}{a table object created from \code{\link[DT]{datatable}()}}

\item{columns}{the indices of the columns to be formatted (can be character,
numeric, logical, or a formula of the form \code{~ V1 + V2}, which is
equivalent to \code{c('V1', 'V2')})}

\item{currency}{the currency symbol}

\item{interval}{put a marker after how many digits of the numbers}

\item{mark}{the marker after every \code{interval} decimals in the numbers}

\item{digits}{the number of decimal places to round to}

\item{dec.mark}{a character to indicate the decimal point}

\item{before}{whether to place the currency symbol before or after the values}

\item{zero.print}{a string to specify how zeros should be formatted.
Useful for when many zero values exist. If \code{NULL}, keeps zero as it is.}

\item{rows}{an integer vector (starting from 1) to specify the only rows
that the style applies to.
By default, it's \code{NULL}, meaning all rows should be formatted. Note,
\code{formatStyle()} doesn't support this argument and you should use
\code{styleRow()} instead. In addition, this only works expected in the
client-side processing mode, i.e., \code{server = FALSE}.}
}
\description{
Format datatable numeric values with commas and no decimal places
}
\examples{
DT_formatNum(DT::datatable(LifeCycleSavings))
}
