% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{DT_hide_col}
\alias{DT_hide_col}
\title{Determine the index of a column name to hide}
\usage{
DT_hide_col(colnames, names)
}
\arguments{
\item{colnames}{\code{chr} of column names}

\item{names}{\code{chr} of columns to hide}
}
\value{
\code{list} with DT option to hide values
}
\description{
Determine the index of a column name to hide
}
