% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{DT_options_update}
\alias{DT_options_update}
\title{Update datatable options}
\usage{
DT_options_update(x, options, hide_cols)
}
\arguments{
\item{x}{\code{datatable}}

\item{options}{\code{(list)} of options to replace}
}
\value{
\code{datatable}
}
\description{
Update datatable options
}
