% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{DT_style_row}
\alias{DT_style_row}
\title{Style rows of a \code{datatable} using source_cols/ID}
\usage{
DT_style_row(
  table,
  columns = 1,
  rows = NULL,
  valueColumns = columns,
  target = c("row", "cell")[2],
  fontWeight = NULL,
  color = DT_style_row_color(rows = rows, values = backgroundColor, default = default),
  default = NULL,
  backgroundColor,
  lineHeight = NULL,
  ...
)
}
\arguments{
\item{table}{\code{datatable}}

\item{columns}{the indices of the columns to be formatted (can be character,
numeric, logical, or a formula of the form \code{~ V1 + V2}, which is
equivalent to \code{c('V1', 'V2')})}

\item{rows}{the Row Indexes (starting from 1) that applies the CSS style. It could
be an integer vector or a list of integer vectors, whose length must be equal to
the length of \code{values}, when \code{values} is not a scalar.}

\item{valueColumns}{indices of the columns from which the cell values are
obtained; this can be different with the \code{columns} argument, e.g. you
may style one column based on the values of a different column}

\item{target}{the target to apply the CSS styles to (the current cell or the
full row)}

\item{fontWeight}{the font weight, e.g. \code{'bold'} and \code{'normal'}}

\item{color}{\code{call} to \code{DT_style_row_color} that uses \code{rows}, \code{values} & \code{default} to impute background and foreground colors based on luminance.}

\item{default}{a string or \code{NULL} used as the the default CSS value
for values other than levels. If \code{NULL}, the CSS value of non-matched
cells will be left unchanged.}

\item{backgroundColor}{the background color of table cells}

\item{...}{
  Arguments passed on to \code{\link[DT:formatCurrency]{DT::formatStyle}}
  \describe{
    \item{\code{background}}{the background of table cells}
  }}
}
\value{
\code{datatable}
}
\description{
Style rows of a \code{datatable} using source_cols/ID
}
