% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{DT_style_row_color}
\alias{DT_style_row_color}
\title{Create a call to \link[DT]{styleRow} for matching color to backgroundColor for contrast}
\usage{
DT_style_row_color(
  rows,
  values,
  default = NULL,
  luminance_threshold = 0.5,
  text_dark = "#333",
  text_light = "#FFF"
)
}
\arguments{
\item{rows}{the Row Indexes (starting from 1) that applies the CSS style. It could
be an integer vector or a list of integer vectors, whose length must be equal to
the length of \code{values}, when \code{values} is not a scalar.}

\item{values}{a vector of CSS values}

\item{default}{a string or \code{NULL} used as the the default CSS value
for values other than levels. If \code{NULL}, the CSS value of non-matched
cells will be left unchanged.}

\item{luminance_threshold}{\code{num} value between 0 1 as the luminance threshold of the background that shifts the color to a contrasting color}

\item{text_dark}{\code{chr} CSS color for dark text}

\item{text_light}{\code{chr} CSS color for light text}
}
\description{
Create a call to \link[DT]{styleRow} for matching color to backgroundColor for contrast
}
