% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{font_color}
\alias{font_color}
\title{Select the black/white font color based on background color value}
\usage{
font_color(values, luminance_threshold = 0.5)
}
\arguments{
\item{values}{\code{chr} hex color values}

\item{luminance_threshold}{\code{dbl} luminance threshold where > this value (brighter background values) will have black font and < this value will have white font}
}
\value{
\code{chr} vector of colors in rgb form
}
\description{
Select the black/white font color based on background color value
}
