% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{render_ran_input}
\alias{render_ran_input}
\title{Create a \code{outputId}_ran input value when a render function rungs}
\usage{
render_ran_input(
  id,
  htmlwidget,
  asis = FALSE,
  session = shiny::getDefaultReactiveDomain(),
  debugger = FALSE
)
}
\arguments{
\item{id}{\code{chr} outputId}

\item{htmlwidget}{\code{htmlwidget} if the object is an htmlwidget, the onRender callback can be used to determine when the widget renders}

\item{asis}{\code{lgl} whether to leave the \code{id} as-is (no namespacing)}

\item{session}{\code{env} The Shiny session object}

\item{debugger}{\code{lgl} Whether to insert a JS debugger statement into the callback}
}
\value{
\code{htmlwidget/none} If \code{htmlwidget} is provided, an \code{htmlwidget} with the \code{onRender} callback added. Otherwise nothing is returned and \link[shinyjs]{runjs} is used internally
}
\description{
Create a \code{outputId}_ran input value when a render function rungs
}
