% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useDTExtend.R
\name{useDTExtend}
\alias{useDTExtend}
\title{Include in the DOM to provide supporting dependencies for this package}
\usage{
useDTExtend()
}
\value{
\code{None} Attaches dependency to the DOM
}
\description{
Include in the DOM to provide supporting dependencies for this package
}
